Crate ethcore_sync

Blockchain sync module Implements ethereum protocol version 63 as specified here: https://github.com/ethereum/wiki/wiki/Ethereum-Wire-Protocol

Structs

Error

The Error type.

EthProtocolInfo

Ethereum protocol info.

EthSync

Ethereum network protocol handler

NetworkConfiguration

Network service configuration

Params

EthSync initialization parameters.

PeerInfo

Peer connection information

PeerNumbers

Numbers of peers (max, min, active).

ServiceConfiguration

Configuration for IPC service.

SyncConfig

Sync configuration

SyncStatus

Syncing status and statistics

TransactionStats

Transaction stats

Enums

ConnectionDirection

Filtered connection direction.

ErrorKind

The kind of an error.

NonReservedPeerMode

Non-reserved peer modes.

PriorityTask

A prioritized tasks run in a specialised timer. Every task should be completed within a hard deadline, if it’s not it’s either cancelled or split into multiple tasks. NOTE These tasks might not complete at all, so anything that happens here should work even if the task is cancelled.

SyncState

Sync state

WarpSync

Determine warp sync status.

Constants

ETH_PROTOCOL

Ethereum sync protocol

PAR_PROTOCOL

OpenEthereum sync protocol

Traits

ConnectionFilter

Connection filter. Each connection is checked against connection_allowed.

ManageNetwork

Trait for managing network

SyncProvider

Current sync status

Functions

validate_node_url

Check if node url is valid